feat: begin support for parsing italic text
This omits important things like italic text split over multiple lines
and other forms of formatting, but it sets out the basic structure of
such.

We also lack an mechanism of escaping the / delimiter, and there's a
corresponding update to the basic should-pass test to reflect this.

Code is also ugly. Way too much indentation. Whatever, it's 23:35.